+++ ostd/specs/mm/page_table/cursor/invariant_preservation_lemmas.rs
@@ -84,11 +84,11 @@ impl<'rcu, C: PageTableConfig> CursorOwner<'rcu, C> {
                 #![trigger cont.children[j]]
                 0 <= j < cont.children.len()
                     && cont.children[j] is Some implies cont.children[j].unwrap().subtree_satisfies(
-            cont.path().push_tail(j as usize), combined) by {
+            cont.path().push_tail(j), combined) by {
                 cont.inv_children_unroll(j);
                 OwnerSubtree::lemma_map_implies_and(
                     cont.children[j].unwrap(),
-                    cont.path().push_tail(j as usize),
+                    cont.path().push_tail(j),
                     f,
                     guard,
                     combined,
@@ -282,12 +282,12 @@ impl<'rcu, C: PageTableConfig> CursorOwner<'rcu, C> {
                 #![trigger cont.children[j]]
                 0 <= j < cont.children.len()
                     && cont.children[j] is Some implies cont.children[j].unwrap().subtree_satisfies(
-            cont.path().push_tail(j as usize), g) by {
+            cont.path().push_tail(j), g) by {
                 cont.inv_children_unroll(j);
                 cont.pt_inv_children_unroll(j);
                 cont.inv_children_rel_unroll(j);
                 let child = cont.children[j].unwrap();
-                let child_path = cont.path().push_tail(j as usize);
+                let child_path = cont.path().push_tail(j);
                 // child.value.path == child_path (inv_children_rel) and
                 // child.value.path.inv() (EntryOwner::inv_base).
                 // L1: pt_inv ⟹ tree-wide path correctness.
